refactor helper methods in rspec
